pyproject.toml
[project]
name = "greenbasket"
version = "0.1.0"
description = ""
authors = [
    {name = "nadezhda",email = "nadezhda13popova@gmail.com"}
]

[tool.poetry.dependencies]
python = ">=3.14,<4.0"
django = ">=6.0.4,<7.0.0"
poetry-core = "^2.3.2"
djangorestframework = "^3.17.1"
python-dotenv = "^1.2.2"
django-apscheduler = "^0.7.0"
pillow = "^12.2.0"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.9.0"
psycopg2 = "^2.9.12"
django-phonenumber-field = "^8.4.0"
phonenumbers = "^9.0.28"
djangorestframework-stubs = "^3.16.9"
django-filter = "^25.2"
django-imagekit = "^6.1.0"
python-slugify = "^8.0.4"
drf-spectacular = "^0.29.0"
gunicorn = "^25.3.0"
whitenoise = "^6.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "black (>=26.3.1,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]
dev = [
    "pre-commit (>=4.6.0,<5.0.0)"
]

[tool.black]
line-length = 119
exclude = """
/(
    .git
  | .hg
  | .mypy_cache
  | .tox
  | .venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | static
  | media
  | node_modules
  | __pycache__
  | manage.py
  | settings.py
  | asgi.py
  | wsgi.py
)/
"""

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
skip_gitignore = true
known_django = ["django"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
